🐛 fix(db, backend): change column types and resolve repository/class errors

+++ backend/src/main/java/com/course/platform/lesson/LessonRepository.java
@@ -2,5 +2,5 @@ package com.course.platform.lesson;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LessonRepository extends JpaRepository<Lesson, Long> {
+public interface LessonRepository extends JpaRepository<Lesson, Integer> {
 }
